+++ tests/test_frontend.py
@@ -12,13 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import contextlib
 import logging
 import os
-import subprocess
-import sys
-from pathlib import Path
-from typing import Any, Callable, cast
+from typing import Any, cast
 
 import pytest
 from streamlit.testing.v1 import AppTest
@@ -27,35 +23,10 @@ logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
-@contextlib.contextmanager
-def cd(new_dir: Path) -> Any:
-    """Changes the current working directory to the given path and restores the old directory on exit."""
-    prev_dir = os.getcwd()
-    os.chdir(new_dir)
-    try:
-        yield
-    finally:
-        os.chdir(prev_dir)
-
-
 @pytest.fixture
-def application(
-    pulumi_up: Any,
-    subprocess_runner: Callable[[list[str]], subprocess.CompletedProcess[str]],
-) -> Any:
-    stack_name = subprocess.check_output(
-        ["pulumi", "stack", "--show-name"],
-        text=True,
-    ).split("\n")[0]
-    with cd(Path("frontend")):
-        subprocess_runner(
-            ["pulumi", "stack", "select", stack_name, "--non-interactive"]
-        )
-        # and ensure we can access `frontend` as if we were running from inside
-        sys.path.append(".")
-        logger.info(subprocess.check_output(["pulumi", "stack", "output"]))
-
-        yield AppTest.from_file("01_connect_and_explore.py", default_timeout=180)
+def application(setup_pulumi_stack: Any) -> Any:
+    """Uses the shared setup_pulumi_stack fixture and returns an AppTest instance"""
+    yield AppTest.from_file("01_connect_and_explore.py", default_timeout=180)
 
 
 @pytest.fixture