+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a bug where snowflake ssh keyfile was not working correctly
 
+### Changed
+
+- Changed the LLMSettings schema for LLM proxy settings instead of Pulumi class
+
 ## [0.1.9] - 2025-04-07
 
 ### Fixed