+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed the incorrect max_completion_length validation in the LLMSettings schema
 
 ### Added
+- Alternative React frontend
 - Allow users to use the app without an API token when it is externally shared
 
 ### Changed