+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.1.10] - 2025-04-17
+
 ### Fixed
 - Fixed a bug where snowflake ssh keyfile was not working correctly
 - Fixed the incorrect max_completion_length validation in the LLMSettings schema